Mode LAN

+++ online.h
@@ -1,7 +1,7 @@
 #define ONLINE_CONFIG_FILE  "config.ini"
 
 void getIPAdress(IPaddress* ip, Uint16* port); // permet de connaitre les informations de connexion de l'adversaire
-UDPsocket initializeConnection(IPaddress* ip, int *port, int* channel);
+UDPsocket initializeConnection(IPaddress* ip, Uint16* port, int* channel);
 void vertexOwnerMessage(char* str, int player, int vertex);
 void removePawnMessage(char* str, int vertex);
 void sendMessage(char* message, int len); // envoie un packet UDP à l'adversaire
